feat: universal hook - fetch, filter, sort
