--- admin/src/assets/styles/components/_ImportExport.scss
@@ -1,6 +1,0 @@
-@use "../common/mixins" as *;
-@use "../common/variables" as *;
-
-.urlslab-importexport {
-	margin: 0 auto;
-}
