+++ admin/src/assets/styles/components/_TableViewHeader.scss
@@ -1,6 +1,22 @@
 @use "../common/variables" as *;
 @use "../common/mixins" as *;
 
-.urlslab-tableView-header {
-	background-color: $white;
+.urlslab-tableView {
+
+	&-header,
+	&-headerBottom {
+		background-color: $white;
+	}
+
+	&-header {
+	}
+
+	&-headerTop {
+		padding: 1.5em 1.625em 0;
+		border-bottom: 1px solid $grey-light;
+	}
+
+	&-headerBottom {
+		padding: 1em 1.625em;
+	}
 }
