+++ admin/src/assets/styles/elements/_Button.scss
@@ -21,7 +21,7 @@ button.urlslab-button,
 	background-color: $grey-light;
 	font-size: 0.8125rem;
 	font-weight: $font-weight-semi;
-	padding: 1em 2em;
+	padding: 0.55em 1em;
 	border-radius: 0.375em;
 	transition: all $transitionTime ease;
 	text-decoration: none;
