+++ admin/src/assets/styles/layouts/_Header.scss
@@ -5,7 +5,6 @@
 	width: 100%;
 	padding: 0.75em 1.625em;
 	background-color: $white;
-	border-bottom: 1px solid $grey-light;
 
 	&-logo {
 		height: 1.32rem;
