+++ admin/src/assets/styles/style.scss
@@ -1,6 +1,7 @@
 @use "./common/variables" as *;
 @use "./common/mixins" as *;
 @forward "./common/global";
+@forward "./common/marginsPaddings";
 @forward "./common/typography";
 @forward "./elements/Button";
 @forward "./layouts/Header";
