+++ admin/src/components/TableViewHeaderBottom.jsx
@@ -2,28 +2,41 @@ import { useI18n } from '@wordpress/react-i18n';
 import { useQueryClient, useMutation } from '@tanstack/react-query';
 import { useCSVReader } from 'react-papaparse';
 import importCsv from '../api/importCsv';
+import { deleteAll } from '../api/deleteTableData';
 
+import { ReactComponent as ImportIcon } from '../assets/images/icon-import.svg';
 import Button from '../elements/Button';
 import ExportCSVButton from '../elements/ExportCSVButton';
-import '../assets/styles/components/_ImportExport.scss';
 
-export default function ImportExport( { children, importOptions, exportOptions } ) {
+export default function TableViewHeaderBottom( { slug, exportOptions } ) {
 	const { __ } = useI18n();
 	const queryClient = useQueryClient();
 	const { CSVReader } = useCSVReader();
 
 	const importData = useMutation( {
 		mutationFn: ( results ) => {
-			return importCsv( `${ importOptions.url }/import`, results.data );
+			return importCsv( `${ slug }/import`, results.data );
 		},
 		onSuccess: () => {
-			queryClient.invalidateQueries( [ importOptions.url ] );
+			queryClient.invalidateQueries( [ slug ] );
+		},
+	} );
+
+	const handleDelete = useMutation( {
+		mutationFn: () => {
+			return deleteAll( slug );
+		},
+		onSuccess: () => {
+			queryClient.invalidateQueries( [ slug ] );
 		},
 	} );
 
 	return (
-		<div className="urlslab-importexport-wrap flex">
-			{ /* <ImportCSVButton options={ importOptions } onClick={ ( data ) => console.log( data ) } /> */ }
+		<div className="urlslab-tableView-headerBottom flex">
+			<Button onClick={ () => handleDelete.mutate() }>{ __( 'Delete All' ) }</Button>
+
+			<ExportCSVButton className="ml-s-tablet" options={ exportOptions } onClick={ ( data ) => console.log( data ) } />
+
 			<CSVReader
 				onUploadAccepted={ ( results ) => {
 					importData.mutate( results );
@@ -38,8 +51,9 @@ export default function ImportExport( { children, importOptions, exportOptions }
 					getRemoveFileProps,
 				} ) => (
 					<>
-						<div>
-							<Button className="active" { ...getRootProps() }>
+						<div className="">
+							<Button className="ml-s-tablet" { ...getRootProps() }>
+								<ImportIcon />
 								{ __( 'Import CSV' ) }
 							</Button>
 							<div>
@@ -49,10 +63,8 @@ export default function ImportExport( { children, importOptions, exportOptions }
 					</>
 				) }
 			</CSVReader>
-
-			<ExportCSVButton options={ exportOptions } onClick={ ( data ) => console.log( data ) } />
 			{
-				queryClient.isMutating() && <h2>Importing</h2>
+				queryClient.isMutating() ? <h2>Importing</h2> : null
 			}
 		</div>
 	);
