+++ admin/src/modules/KeywordsLinks.jsx
@@ -2,24 +2,29 @@ import { useState, Suspense, lazy } from 'react';
 import { useI18n } from '@wordpress/react-i18n';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import TableViewHeader from '../components/TableViewHeader';
-import KeywordsTable from '../tables/KeywordsTable';
 
 export default function KeywordLinks( { moduleId } ) {
 	const { __ } = useI18n();
+	const slug = 'keyword';
 	const [ activeSection, setActiveSection ] = useState( 'overview' );
 
 	const tableMenu = new Map( [
-		[ 'keyword', __( 'Keywords Table' ) ],
+		[ slug, __( 'Keywords Table' ) ],
 	] );
 
 	const SettingsModule = lazy( () => import( `../modules/Settings.jsx` ) );
+	const KeywordsTable = lazy( () => import( `../tables/KeywordsTable.jsx` ) );
 
 	return (
 		<div className="urlslab-tableView">
-			<TableViewHeader tableMenu={ tableMenu } activeMenu={ ( activemenu ) => setActiveSection( activemenu ) } />
+			<TableViewHeader
+				tableMenu={ tableMenu }
+				activeMenu={ ( activemenu ) => setActiveSection( activemenu ) } />
 			{
-				activeSection === 'keyword' &&
-				<KeywordsTable slug="keyword" />
+				activeSection === slug &&
+				<Suspense>
+					<KeywordsTable slug={ slug } />
+				</Suspense>
 			}
 			{
 				activeSection === 'settings' &&
