+++ admin/src/modules/LinkEnhancer.jsx
@@ -2,7 +2,6 @@ import { useState, Suspense, lazy } from 'react';
 import { useI18n } from '@wordpress/react-i18n';
 
 import TableViewHeader from '../components/TableViewHeader';
-import LinkManagerTable from '../tables/LinkManagerTable';
 
 export default function LinkEnhancer( { moduleId } ) {
 	const { __ } = useI18n();
@@ -13,13 +12,16 @@ export default function LinkEnhancer( { moduleId } ) {
 	] );
 
 	const SettingsModule = lazy( () => import( `../modules/Settings.jsx` ) );
+	const LinkManagerTable = lazy( () => import( `../tables/LinkManagerTable.jsx` ) );
 
 	return (
 		<div className="urlslab-tableView">
 			<TableViewHeader tableMenu={ tableMenu } activeMenu={ ( activemenu ) => setActiveSection( activemenu ) } />
 			{
 				activeSection === 'url' &&
-				<LinkManagerTable />
+				<Suspense>
+					<LinkManagerTable />
+				</Suspense>
 			}
 			{
 				activeSection === 'settings' &&
