+++ admin/src/modules/RelatedResources.jsx
@@ -2,7 +2,6 @@ import { useState, Suspense, lazy } from 'react';
 import { useI18n } from '@wordpress/react-i18n';
 
 import TableViewHeader from '../components/TableViewHeader';
-import URLRelationTable from '../tables/URLRelationTable';
 
 export default function RelatedResources( { moduleId } ) {
 	const { __ } = useI18n();
@@ -13,13 +12,16 @@ export default function RelatedResources( { moduleId } ) {
 	] );
 
 	const SettingsModule = lazy( () => import( `../modules/Settings.jsx` ) );
+	const URLRelationTable = lazy( () => import( `../tables/URLRelationTable.jsx` ) );
 
 	return (
 		<div className="urlslab-tableView">
 			<TableViewHeader tableMenu={ tableMenu } activeMenu={ ( activemenu ) => setActiveSection( activemenu ) } />
 			{
 				activeSection === 'url-relation' &&
-				<URLRelationTable />
+				<Suspense>
+					<URLRelationTable />
+				</Suspense>
 			}
 			{
 				activeSection === 'settings' &&
