+++ admin/src/tables/LinkManagerTable.jsx
@@ -1,55 +1,76 @@
-import { useEffect } from 'react';
-import { useInfiniteQuery } from '@tanstack/react-query';
-import { useInView } from 'react-intersection-observer';
-import { fetchData } from '../api/fetching';
+import { useState } from 'react';
+import { useI18n } from '@wordpress/react-i18n';
+import { createColumnHelper } from '@tanstack/react-table';
 
-import Columns from './tableColumns/LinkManagerTable';
-import Table from '../components/TableComponent';
+import useInfiniteFetch from '../hooks/useInfiniteFetch';
+import { handleInput, handleSelected } from '../constants/tableFunctions';
+
+import SortMenu from '../elements/SortMenu';
+import Checkbox from '../elements/Checkbox';
 
 import Loader from '../components/Loader';
 
+import Table from '../components/TableComponent';
+import TableViewHeaderBottom from '../components/TableViewHeaderBottom';
+
 export default function LinkManagerTable() {
-	const { ref, inView } = useInView();
-	const maxRows = 50;
+	const { __ } = useI18n();
+	const columnHelper = createColumnHelper();
+	const [ currentUrl, setUrl ] = useState();
 
 	const {
 		data,
 		status,
 		isSuccess,
 		isFetchingNextPage,
-		fetchNextPage,
 		hasNextPage,
-	} = useInfiniteQuery( {
-		queryKey: [ 'url' ],
-		queryFn: ( { pageParam = 0 } ) => {
-			return fetchData( `url?from_urlId=${ pageParam }&rows_per_page=${ maxRows }` );
-		},
-		getNextPageParam: ( allRows ) => {
-			if ( allRows.length < maxRows ) {
-				return undefined;
-			}
-			const lastRowId = allRows[ allRows?.length - 1 ]?.urlId ?? undefined;
-			return lastRowId;
-		},
-		keepPreviousData: true,
-		refetchOnWindowFocus: false,
-		cacheTime: Infinity,
-		staleTime: Infinity,
-	}
-	);
+		ref,
+	} = useInfiniteFetch( { key: 'url', url: currentUrl, pageId: 'urlId' } );
 
-	useEffect( () => {
-		if ( inView ) {
-			fetchNextPage();
-		}
-	}, [ inView, fetchNextPage ] );
+	const statusTypes = {
+		N: __( 'New' ),
+		A: __( 'Available' ),
+		P: __( 'Processing' ),
+		D: __( 'Disabled' ),
+	};
+
+	const columns = [
+		columnHelper.accessor( 'check', {
+			className: 'checkbox',
+			cell: ( cell ) => <Checkbox checked={ cell.row.getIsSelected() } onChange={ ( val ) => {
+				handleSelected( val, cell );
+			} } />,
+			header: () => __( '' ),
+		} ),
+		columnHelper.accessor( 'urlTitle', {
+			header: () => __( 'URL Title' ),
+		} ),
+		columnHelper.accessor( 'urlName', {
+			header: () => __( 'URL' ),
+		} ),
+		columnHelper?.accessor( 'urlMetaDescription', {
+			header: () => __( 'URL Description' ),
+		} ),
+		columnHelper?.accessor( 'status', {
+			cell: ( cell ) => <SortMenu
+				items={ statusTypes }
+				name={ cell.column.id }
+				checkedId={ cell.getValue() }
+				onChange={ ( val ) => handleInput( val, cell ) } />,
+			className: 'youtube-status',
+			header: () => __( 'Status' ),
+		} ),
+		columnHelper.accessor( 'updateStatusDate', {
+			header: () => __( 'Status Date' ),
+		} ),
+	];
 
 	if ( status === 'loading' ) {
 		return <Loader />;
 	}
 
 	return (
-		<Table className="fadeInto" columns={ Columns() }
+		<Table className="fadeInto" columns={ columns }
 			data={
 				isSuccess && data?.pages?.flatMap( ( page ) => page ?? [] )
 			}
